refactor: co-locate useAnimationController hook with AnimationController class
Move useAnimationController hook from src/hooks/ to src/features/visuals/map/animations/
to co-locate it with the AnimationController class it wraps.

This improves code organization by:
- Keeping animation control logic together in one feature folder
- Making the relationship between hook and class explicit
- Following domain-driven design principles

The hook is now exported from the animations index for easy importing.
