+++ src/features/visuals/map/animations/useAnimationController.ts
